feat: AudioRecord에 서버 연결 과정 통합. 보내는 오디오 파형 확인할 수 있게 구현 #34
AudioRecord에 서버 연결 후 전송 과정을 통합했습니다. 보내는 오디오의
파형도 확인 가능합니다. 오디오 처리와 서버 전송과정 함수들을 작게
분리했습니다.
